TEDEFO-1415 split of controllers and urls.
